working on delete confirm

+++ src/components/PostsList/Post/Post.js
@@ -5,7 +5,8 @@ import { Link } from 'react-router-dom';
 import UserContext from '../../../contexts/UserContext';
 import Button from '../../Button/Button';
 import PostsApiService from '../../../services/posts-api-service';
-import DeleteConfirmation from '../../DeleteConfirmation/DeleteConfirmation'
+import Confirm from '../../Confirm/Confirm';
+import '@reach/dialog/styles.css';
 
 export default class Post extends React.Component {
   state = {
@@ -25,18 +26,23 @@ export default class Post extends React.Component {
 
     let deleteButton =
       this.context.user.id === post.user_id ? (
-        <Button
-          onClick={(e) => {
-            e.preventDefault();
-            PostsApiService.deletePost(post.id);
-            this.props.deletePost(post.id);
-          }}
-          type="delete"
-          title="Delete"
-          className={styles.deletePostButton}
-          id={styles.deletePostButton}>
-          X
-        </Button>
+        <Confirm title="Confirm" description="Are you sure?">
+          {(confirm) => (
+            <form
+              onSubmit={confirm(() => {
+                PostsApiService.deletePost(post.id);
+                this.props.deletePost(post.id);
+              })}>
+              <Button
+                type="delete"
+                title="Delete"
+                className={styles.deletePostButton}
+                id={styles.deletePostButton}>
+                X
+              </Button>
+            </form>
+          )}
+        </Confirm>
       ) : (
         <Button
           type="delete"
@@ -48,41 +54,36 @@ export default class Post extends React.Component {
       );
 
     return (
-      <Link to={`/post/${post.id}`}>
-        <li
-          className={styles.Post}
-          onMouseEnter={this.handleListHoverOn}
-          onMouseLeave={this.handleListHoverOff}>
-          {/* <div className={styles.typeCircle}></div> */}
-          <span
-            className={cx(
-              styles.postTitle,
-              styles.postEl,
+      <li
+        className={styles.Post}
+        onMouseEnter={this.handleListHoverOn}
+        onMouseLeave={this.handleListHoverOff}>
+        {/* <div className={styles.typeCircle}></div> */}
+        <span className={cx(styles.PostTitle, styles.postEl)}>
+          <Link
+            className={
               post.type === 'offer' ? styles.offerStyle : styles.requestStyle
-            )}>
+            }
+            to={`/post/${post.id}`}>
             {title}
-            {this.state.hover && deleteButton}
-          </span>
-          {/* <DeleteConfirmation>  <p>you want to delete???</p>         </DeleteConfirmation> */}
+          </Link>
+          {this.state.hover && deleteButton}
+        </span>
 
-
-          <span className={cx(styles.postType, styles.postEl)}>
-            {post.type}
-          </span>
-          <span className={cx(styles.postComments, styles.postEl)}>
-            {post.comments}
-          </span>
-          <span className={cx(styles.postUserName, styles.postEl)}>
-            {post.name}
-          </span>
-          <span className={cx(styles.postDateFull, styles.postEl)}>
-            {post.date_modified.slice(0, 10)}
-          </span>
-          <span className={cx(styles.postDateTrim, styles.postEl)}>
-            {post.date_modified.slice(5, 10)}
-          </span>
-        </li>
-      </Link>
+        <span className={cx(styles.postType, styles.postEl)}>{post.type}</span>
+        <span className={cx(styles.postComments, styles.postEl)}>
+          {post.comments}
+        </span>
+        <span className={cx(styles.postUserName, styles.postEl)}>
+          {post.name}
+        </span>
+        <span className={cx(styles.postDateFull, styles.postEl)}>
+          {post.date_modified.slice(0, 10)}
+        </span>
+        <span className={cx(styles.postDateTrim, styles.postEl)}>
+          {post.date_modified.slice(5, 10)}
+        </span>
+      </li>
     );
   }
 }
